Refactor by WayangRules TODO split into multiple classes
