+++ wayang-api/wayang-api-sql/src/main/java/org/apache/wayang/api/sql/calcite/rules/WayangRules.java
@@ -40,12 +40,10 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import java.util.ArrayList;
 import java.util.List;
 
-//TODO: split into multiple classes
 public class WayangRules {
 
     private WayangRules(){
     }
-
 
     public static final RelOptRule WAYANG_JOIN_RULE = new WayangJoinRule(WayangJoinRule.DEFAULT_CONFIG);
     public static final RelOptRule WAYANG_PROJECT_RULE = new WayangProjectRule(WayangProjectRule.DEFAULT_CONFIG);
@@ -54,157 +52,4 @@ public class WayangRules {
     public static final RelOptRule WAYANG_TABLESCAN_ENUMERABLE_RULE =
             new WayangTableScanRule(WayangTableScanRule.ENUMERABLE_CONFIG);
     public static final RelOptRule WAYANG_AGGREGATE_RULE =  new WayangAggregateRule(WayangAggregateRule.DEFAULT_CONFIG);
-
-
-    private static class WayangProjectRule extends ConverterRule {
-
-        public static final Config DEFAULT_CONFIG = Config.INSTANCE
-                .withConversion(LogicalProject.class,
-                        Convention.NONE, WayangConvention.INSTANCE,
-                        "WayangProjectRule")
-                .withRuleFactory(WayangProjectRule::new);
-
-
-        protected WayangProjectRule(Config config) {
-            super(config);
-        }
-
-        public RelNode convert(RelNode rel) {
-            final LogicalProject project = (LogicalProject) rel;
-            return new WayangProject(
-                    project.getCluster(),
-                    project.getTraitSet().replace(WayangConvention.INSTANCE),
-                    convert(project.getInput(), project.getInput().getTraitSet()
-                            .replace(WayangConvention.INSTANCE)),
-                    project.getProjects(),
-                    project.getRowType());
-        }
-    }
-
-
-    private static class WayangFilterRule extends ConverterRule {
-
-        public static final Config DEFAULT_CONFIG = Config.INSTANCE
-                .withConversion(LogicalFilter.class,
-                        Convention.NONE, WayangConvention.INSTANCE,
-                        "WayangFilterRule")
-                .withRuleFactory(WayangFilterRule::new);
-
-
-        protected WayangFilterRule(Config config) {
-            super(config);
-        }
-
-        @Override
-        public RelNode convert(RelNode rel) {
-            final LogicalFilter filter = (LogicalFilter) rel;
-            return new WayangFilter(
-                    rel.getCluster(),
-                    rel.getTraitSet().replace(WayangConvention.INSTANCE),
-                    convert(filter.getInput(), filter.getInput().getTraitSet().
-                            replace(WayangConvention.INSTANCE)),
-                    filter.getCondition());
-        }
-
-    }
-
-    private static class WayangTableScanRule extends ConverterRule {
-
-        public static final Config DEFAULT_CONFIG = Config.INSTANCE
-                .withConversion(LogicalTableScan.class,
-                        Convention.NONE, WayangConvention.INSTANCE,
-                        "WayangTableScanRule")
-                .withRuleFactory(WayangTableScanRule::new);
-
-        public static final Config ENUMERABLE_CONFIG = Config.INSTANCE
-                .withConversion(TableScan.class,
-                        EnumerableConvention.INSTANCE, WayangConvention.INSTANCE,
-                        "WayangTableScanRule1")
-                .withRuleFactory(WayangTableScanRule::new);
-
-        protected WayangTableScanRule(Config config) {
-            super(config);
-        }
-
-        @Override
-        public @Nullable RelNode convert(RelNode relNode) {
-
-            TableScan scan = (TableScan) relNode;
-            final RelOptTable relOptTable = scan.getTable();
-
-            /**
-             * This is quick hack to prevent volcano from merging projects on to TableScans
-             * TODO: a cleaner way to handle this
-             */
-            if(relOptTable.getRowType() == scan.getRowType()) {
-                return WayangTableScan.create(scan.getCluster(), relOptTable);
-            }
-            return null;
-        }
-    }
-
-
-    private static class WayangJoinRule extends ConverterRule {
-
-        public static final Config DEFAULT_CONFIG = Config.INSTANCE
-                .withConversion(LogicalJoin.class, Convention.NONE,
-                        WayangConvention.INSTANCE, "WayangJoinRule")
-                .withRuleFactory(WayangJoinRule::new);
-
-        protected WayangJoinRule(Config config) {
-            super(config);
-        }
-
-        @Override
-        public @Nullable RelNode convert(RelNode relNode) {
-            LogicalJoin join = (LogicalJoin) relNode;
-            List<RelNode> newInputs = new ArrayList<>();
-            for(RelNode input : join.getInputs()) {
-                if(!(input.getConvention() instanceof WayangConvention)) {
-                    input = convert(input, input.getTraitSet().replace(WayangConvention.INSTANCE));
-                }
-                newInputs.add(input);
-            }
-
-            return new WayangJoin(
-                    join.getCluster(),
-                    join.getTraitSet().replace(WayangConvention.INSTANCE),
-                    newInputs.get(0),
-                    newInputs.get(1),
-                    join.getCondition(),
-                    join.getVariablesSet(),
-                    join.getJoinType()
-            );
-        }
-    }
-    private static class WayangAggregateRule extends ConverterRule {
-
-        public static final Config DEFAULT_CONFIG = Config.INSTANCE
-                .withConversion(LogicalAggregate.class,
-                        Convention.NONE, WayangConvention.INSTANCE,
-                        "WayangAggregateRule")
-                .withRuleFactory(WayangAggregateRule::new);
-
-        protected WayangAggregateRule(Config config) {
-            super(config);
-        }
-
-        @Override
-        public @Nullable RelNode convert(RelNode relNode) {
-            LogicalAggregate aggregate = (LogicalAggregate) relNode;
-            RelNode input = convert(aggregate.getInput(), aggregate.getInput().getTraitSet().replace(WayangConvention.INSTANCE));
-
-            return new WayangAggregate(
-                    aggregate.getCluster(),
-                    aggregate.getTraitSet().replace(WayangConvention.INSTANCE),
-                    aggregate.getHints(),
-                    input,
-                    aggregate.getGroupSet(),
-                    aggregate.getGroupSets(),
-                    aggregate.getAggCallList()
-            );
-        }
-    }
-
-
 }
